Weather Fragment Current location functionality complete - The main weather fragment now populates a current, 24 hour, and a 10 day forecast based on the devices current location. - The CurrentWeatherFragment is to be used on the dynamic home page. -TODO: implement zipcode and map functionality for weather locations. - implement save/load locations from shared preferences

+++ app/src/main/java/tcss450/uw/edu/chapp/weather/WeatherHourContent.java
@@ -46,10 +46,10 @@ public class WeatherHourContent {
     public static class WeatherHourItem {
         public final String id;
         public final String time;
-        public final int temp;
+        public final double temp;
         public final String icon;
 
-        public WeatherHourItem(String id, String time, int temp, String icon) {
+        public WeatherHourItem(String id, String time, double temp, String icon) {
             this.id = id;
             this.time = time;
             this.temp = temp;
